requirements.txt
azure-ai-ml>=1.14.0
azure-identity>=1.15.0
